move conflict detector into the mysql sink
