+++ downstreamadapter/sink/mysql/causality/conflict_detector.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package conflictdetector
+package causality
 
 import (
 	"context"
