+++ downstreamadapter/sink/mysql/causality/node.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package conflictdetector
+package causality
 
 import (
 	"sync"
